Fix typo in the project name
