+++ noql-backend/src/main/java/com/janbabak/noqlbackend/service/api/QueryApi.java
@@ -1,0 +1,5 @@
+package com.janbabak.noqlbackend.service.api;
+
+public interface QueryApi {
+    String queryModel(String query) throws Exception;
+}
