+++ noql-backend/src/test/java/com/janbabak/noqlbackend/NoQLBackendApplicationTests.java
@@ -1,0 +1,13 @@
+package com.janbabak.noqlbackend;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NoqlBackendApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
