--- noql-bakend/src/main/java/com/janbabak/noqlbakend/service/api/QueryApi.java
@@ -1,5 +1,0 @@
-package com.janbabak.noqlbakend.service.api;
-
-public interface QueryApi {
-    String queryModel(String query) throws Exception;
-}
