--- noql-bakend/src/test/java/com/janbabak/noqlbakend/NoqlBakendApplicationTests.java
@@ -1,13 +1,0 @@
-package com.janbabak.noqlbakend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NoqlBakendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
